+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Stories
 * [UINV-118](https://issues.folio.org/browse/UINV-117) Allow user to edit voucher information
+* [UINV-103](https://issues.folio.org/browse/UINV-103) Add voucher settings for a given Batch group
 * [UINV-118](https://issues.folio.org/browse/UINV-118) TECH-DEBT refactor Invoice list to not use SearchAndSort
 * [UINV-122](https://issues.folio.org/browse/UINV-122) Add "always show" preset adjustments to invoice when creating new invoice
 * [UINV-101](https://issues.folio.org/browse/UINV-101) Invoice settings: Batch groupsCreate settings form for Batch group settings (Multiple Batch groups)